// src/types.ts
export type Entry = {
    id: number;
    date: string| null;
    intensity: number | null;
    duration: number | null;
    total_hours_of_migraine: number | null;
    episode_date: string | null;
    estimated_onset_time: string | null;
    estimated_onset_period: 'AM' | 'PM' | null;
    estimated_ended_time: string | null;
    estimated_ended_period: 'AM' | 'PM' | null;
    recent_duration_of_sleep: string | null;
    site_of_pain_front: boolean | null;
    site_of_pain_back: boolean | null;
    site_of_pain_left: boolean | null;
    site_of_pain_right: boolean | null;
    site_of_pain_top: boolean | null;
    headache_severity: 'Mild' | 'Moderate' | 'Severe' | 'Extreme' | null;
    weather: string | null;
    temperature_high: number | null;
    temperature_low: number | null;
    hydration_oz: number | null;
    symptoms_throbbing: boolean | null;
    symptoms_burning: boolean | null;
    symptoms_dull_ache: boolean | null;
    symptoms_knife_like: boolean | null;
    symptoms_nausea: boolean | null;
    symptoms_light_sensitivity: boolean | null;
    symptoms_pressure: boolean | null;
    symptoms_aura: boolean | null;
    symptoms_tight_band: boolean | null;
    symptoms_neck_ache: boolean | null;
    warning_vision: boolean | null;
    warning_numbness: boolean | null;
    warning_aching_neck: boolean | null;
    warning_other: string | null;
    factors_brought_on: string | null;
    factors_relieve: string | null;
}